+++ .eslintrc.js
@@ -95,6 +95,8 @@ module.exports = {
         '@typescript-eslint/no-unnecessary-condition': 'off',
         '@typescript-eslint/consistent-type-imports': ['error', { prefer: 'type-imports' }],
         '@typescript-eslint/unbound-method': 'off',
+        '@typescript-eslint/sort-type-union-intersection-members': 'off',
+        '@typescript-eslint/no-non-null-assertion': 'off',
       },
     },
     {