+++ packages/blade/src/components/BladeProvider/__tests__/paymentLightTheme/paymentLightTheme.web.ts
@@ -54,32 +54,32 @@ const paymentLightTheme: Theme = {
           highContrast: globalColors.neutral.blueGrayLight[700],
         },
       },
-      action: {
-        icon: {
-          link: {
-            default: {
-              lowContrast: globalColors.neutral.blueGrayLight[500],
-              highContrast: globalColors.neutral.blueGrayLight[300],
-            },
-            hover: {
-              lowContrast: globalColors.neutral.blueGrayLight[800],
-              highContrast: globalColors.neutral.blueGrayLight[0],
-            },
-            focus: {
-              lowContrast: globalColors.neutral.blueGrayLight[800],
-              highContrast: globalColors.neutral.blueGrayLight[0],
-            },
-            active: {
-              lowContrast: globalColors.neutral.blueGrayLight[800],
-              highContrast: globalColors.neutral.blueGrayLight[0],
-            },
-            disabled: {
-              lowContrast: globalColors.neutral.blueGrayLight[300],
-              highContrast: globalColors.neutral.blueGrayLight.a200,
-            },
-          },
-        },
-      },
+      // action: {
+      //   icon: {
+      //     link: {
+      //       default: {
+      //         lowContrast: globalColors.neutral.blueGrayLight[500],
+      //         highContrast: globalColors.neutral.blueGrayLight[300],
+      //       },
+      //       hover: {
+      //         lowContrast: globalColors.neutral.blueGrayLight[800],
+      //         highContrast: globalColors.neutral.blueGrayLight[0],
+      //       },
+      //       focus: {
+      //         lowContrast: globalColors.neutral.blueGrayLight[800],
+      //         highContrast: globalColors.neutral.blueGrayLight[0],
+      //       },
+      //       active: {
+      //         lowContrast: globalColors.neutral.blueGrayLight[800],
+      //         highContrast: globalColors.neutral.blueGrayLight[0],
+      //       },
+      //       disabled: {
+      //         lowContrast: globalColors.neutral.blueGrayLight[300],
+      //         highContrast: globalColors.neutral.blueGrayLight.a200,
+      //       },
+      //     },
+      //   },
+      // },
     },
     overlay: {
       background: globalColors.neutral.blueGrayLight.a200,
@@ -93,7 +93,8 @@ const paymentLightTheme: Theme = {
         700: globalColors.chromatic.azure[700],
       },
       gray: {
-        300: globalColors.neutral.blueGrayLight[50],
+        200: globalColors.neutral.blueGrayLight[50],
+        300: globalColors.neutral.blueGrayLight[100],
         400: globalColors.neutral.blueGrayLight[300],
         500: globalColors.neutral.blueGrayLight[500],
         600: globalColors.neutral.blueGrayLight[600],
@@ -948,14 +949,14 @@ const paymentLightTheme: Theme = {
           hover: globalColors.chromatic.azure[500],
           focus: globalColors.chromatic.azure[500],
           active: globalColors.chromatic.azure[500],
-          disabled: globalColors.neutral.blueGrayLight[600],
+          disabled: globalColors.neutral.blueGrayLight[400],
         },
         tertiary: {
-          default: globalColors.neutral.blueGrayLight[400],
-          hover: globalColors.neutral.blueGrayLight[400],
-          focus: globalColors.neutral.blueGrayLight[400],
-          active: globalColors.neutral.blueGrayLight[400],
-          disabled: globalColors.neutral.blueGrayLight[400],
+          default: globalColors.neutral.blueGrayLight[300],
+          hover: globalColors.neutral.blueGrayLight[300],
+          focus: globalColors.neutral.blueGrayLight[300],
+          active: globalColors.neutral.blueGrayLight[300],
+          disabled: globalColors.neutral.blueGrayLight[300],
         },
       },
       text: {
@@ -967,18 +968,18 @@ const paymentLightTheme: Theme = {
           disabled: globalColors.neutral.blueGrayLight[600],
         },
         secondary: {
-          default: globalColors.chromatic.azure[600],
-          hover: globalColors.chromatic.azure[600],
-          focus: globalColors.chromatic.azure[600],
-          active: globalColors.chromatic.azure[600],
-          disabled: globalColors.neutral.blueGrayLight[600],
+          default: globalColors.chromatic.azure[500],
+          hover: globalColors.chromatic.azure[500],
+          focus: globalColors.chromatic.azure[500],
+          active: globalColors.chromatic.azure[500],
+          disabled: globalColors.neutral.blueGrayLight[400],
         },
         tertiary: {
           default: globalColors.neutral.blueGrayLight[1000],
           hover: globalColors.neutral.blueGrayLight[1000],
           focus: globalColors.neutral.blueGrayLight[1000],
           active: globalColors.neutral.blueGrayLight[1000],
-          disabled: globalColors.neutral.blueGrayLight[500],
+          disabled: globalColors.neutral.blueGrayLight[400],
         },
         link: {
           default: globalColors.chromatic.azure[500],
@@ -997,18 +998,18 @@ const paymentLightTheme: Theme = {
           disabled: globalColors.neutral.blueGrayLight[600],
         },
         secondary: {
-          default: globalColors.chromatic.azure[600],
-          hover: globalColors.chromatic.azure[600],
-          focus: globalColors.chromatic.azure[600],
-          active: globalColors.chromatic.azure[600],
-          disabled: globalColors.neutral.blueGrayLight[600],
+          default: globalColors.chromatic.azure[500],
+          hover: globalColors.chromatic.azure[500],
+          focus: globalColors.chromatic.azure[500],
+          active: globalColors.chromatic.azure[500],
+          disabled: globalColors.neutral.blueGrayLight[400],
         },
         tertiary: {
           default: globalColors.neutral.blueGrayLight[1000],
           hover: globalColors.neutral.blueGrayLight[1000],
           focus: globalColors.neutral.blueGrayLight[1000],
           active: globalColors.neutral.blueGrayLight[1000],
-          disabled: globalColors.neutral.blueGrayLight[500],
+          disabled: globalColors.neutral.blueGrayLight[400],
         },
         link: {
           default: globalColors.chromatic.azure[500],