+++ packages/blade/src/components/Icons/ChevronDownIcon/ChevronDownIcon.web.test.tsx
@@ -1,0 +1,11 @@
+import renderWithTheme from '../../../_helpers/testing/renderWithTheme.web';
+import ChevronDownIcon from '.';
+
+describe('<ChevronDownIcon />', () => {
+  it('should render ChevronDownIcon', () => {
+    const { container } = renderWithTheme(
+      <ChevronDownIcon color="feedback.icon.neutral.lowContrast" size="large" />,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});