+++ packages/blade/src/components/Icons/_Svg/Rect/Rect.d.ts
@@ -1,6 +1,7 @@
 export type RectProps = {
   width: string;
   height: string;
+  fill?: string;
   rx?: string;
   ry?: string;
   x?: string;