+++ packages/blade/src/components/Typography/Text/Text.stories.tsx
@@ -1,8 +1,52 @@
-/* eslint-disable react/display-name */
 import type { ComponentStory, Meta } from '@storybook/react';
-import { Title, Subtitle, Primary, Props, Stories, PRIMARY_STORY } from '@storybook/addon-docs';
+import { Title, Subtitle, Primary, ArgsTable, Stories, PRIMARY_STORY } from '@storybook/addon-docs';
+import { Highlight, Link } from '@storybook/design-system';
+import type { ReactElement } from 'react';
+import useMakeFigmaURL from '../../../_helpers/storybook/useMakeFigmaURL';
 import type { TextProps } from './';
 import TextComponent from './';
+
+const Page = (): ReactElement => {
+  const figmaURL = useMakeFigmaURL([
+    {
+      themeTokenName: 'paymentTheme',
+      lightModeURL:
+        'https://www.figma.com/file/jubmQL9Z8V7881ayUD95ps/Blade---Payment-Light?node-id=11770%3A147139',
+      darkModeURL:
+        'https://www.figma.com/file/jubmQL9Z8V7881ayUD95ps/Blade---Payment-Light?node-id=11770%3A147139',
+    },
+    {
+      themeTokenName: 'bankingTheme',
+      lightModeURL:
+        'https://www.figma.com/file/sAdplk2uYnI2ILnDKUxycW/Blade---Banking-Dark?node-id=10344%3A189844',
+      darkModeURL:
+        'https://www.figma.com/file/sAdplk2uYnI2ILnDKUxycW/Blade---Banking-Dark?node-id=10344%3A189844',
+    },
+  ]);
+
+  return (
+    <>
+      <Title />
+      <Subtitle>
+        The Text component is used to display main content of the page. It is often clubbed with
+        Title or Heading to display content in a hierarchical structure. It applies responsive
+        styles automatically based on the device it is being rendered on.
+      </Subtitle>
+      <Link withArrow={true} href={figmaURL} target="_blank">
+        View in Figma
+      </Link>
+      <br />
+      <br />
+      <Title>Usage</Title>
+      <Highlight language="tsx">{`import { Text } from '@razorpay/blade/components' \nimport type { TextProps } from '@razorpay/blade/components'`}</Highlight>
+      <Title>Example</Title>
+      <Primary />
+      <Title>Properties</Title>
+      <ArgsTable story={PRIMARY_STORY} />
+      <Stories />
+    </>
+  );
+};
 
 const TextStoryMeta: Meta<TextProps<{ variant: 'body' | 'caption' }>> = {
   title: 'Components/Typography/Text',
@@ -14,24 +58,11 @@ const TextStoryMeta: Meta<TextProps<{ variant: 'body' | 'caption' }>> = {
     children:
       'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis, ultricies nec, pellentesque eu, pretium quis, sem. Nulla consequat massa quis enim. Donec pede justo, fringilla vel, aliquet nec, vulputate eget, arcu. In enim justo, rhoncus ut, imperdiet a, venenatis vitae, justo. Nullam dictum felis eu pede mollis pretium. Integer tincidunt. Cras dapibus. Vivamus elementum semper nisi. Aenean vulputate eleifend tellus. Aenean leo ligula, porttitor eu, consequat vitae, eleifend ac, enim. Aliquam lorem ante, dapibus in, viverra quis, feugiat a, tellus. Phasellus viverra nulla ut metus varius laoreet. Quisque rutrum. Aenean imperdiet. Etiam ultricies nisi vel augue. Curabitur ullamcorper ultricies nisi. Nam eget dui. Etiam rhoncus. Maecenas tempus, tellus eget condimentum rhoncus, sem quam semper libero, sit amet adipiscing sem neque sed ipsum. Nam quam nunc, blandit vel, luctus pulvinar, hendrerit id, lorem. Maecenas nec odio et ante tincidunt tempus. Donec vitae sapien ut libero venenatis faucibus. Nullam quis ante. Etiam sit amet orci eget eros faucibus tincidunt. Duis leo. Sed fringilla mauris sit amet nibh. Donec sodales sagittis magna. Sed consequat, leo eget bibendum sodales, augue velit cursus nunc',
     truncateAfterLines: 3,
+    contrast: 'low',
   },
   parameters: {
     docs: {
-      page: () => (
-        <>
-          <Title />
-          <Subtitle>
-            The Text component is used to display main content of the page. It is often clubbed with
-            Title or Heading to display content in a hierarchical structure. It applies responsive
-            styles automatically based on the device it is being rendered on.
-          </Subtitle>
-          <Title>Example</Title>
-          <Primary />
-          <Title>Properties</Title>
-          <Props story={PRIMARY_STORY} />
-          <Stories />
-        </>
-      ),
+      page: () => <Page />,
     },
   },
 };